google_speech_to_text/requirements.txt
python-dotenv
pydub
requests
